requirements.txt
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-crontab==0.7.1
django-rest-knox==4.2.0